In order to use Wirespec in kotlin multiplatform no JVM dependencies are allowed change from javaType to KType in kotlin shared
