--- examples/spring-boot-openapi-maven-plugin/src/main/kotlin/community/flock/wirespec/examples/open_api_app/java/JavaPetstoreController.kt
@@ -1,38 +1,0 @@
-package community.flock.wirespec.examples.open_api_app.java
-
-import community.flock.wirespec.generated.java.v3.AddPetEndpoint
-import community.flock.wirespec.generated.java.v3.FindPetsByStatusEndpoint
-import community.flock.wirespec.generated.java.v3.FindPetsByStatusParameterStatus
-import community.flock.wirespec.generated.java.v3.Pet
-import org.springframework.web.bind.annotation.*
-import java.util.*
-import kotlin.jvm.optionals.getOrNull
-
-@RestController
-@RequestMapping("/java/pets")
-class JavaPetstoreController(
-    private val javaPetstoreClient: JavaPetstoreClient
-) {
-
-    @GetMapping
-    suspend fun addPet(): Optional<Long>? {
-        val pet = Pet(Optional.empty(), "Petje", Optional.empty(), emptyList(), Optional.empty(), Optional.empty())
-        val req = AddPetEndpoint.RequestApplicationJson(pet)
-        return when (val res = javaPetstoreClient.addPet(req).get()) {
-            is AddPetEndpoint.Response200ApplicationJson -> res.content?.body?.id
-            else -> error("No response")
-        }
-    }
-
-    @PostMapping
-    suspend fun create(@RequestBody pet: Pet): List<Long?> {
-        val req = FindPetsByStatusEndpoint.RequestUnit(Optional.of(FindPetsByStatusParameterStatus.available))
-        return when (val res = javaPetstoreClient.findPetsByStatus(req).get()) {
-            is FindPetsByStatusEndpoint.Response200ApplicationJson -> res.content?.body?.mapNotNull { it.id.getOrNull() }
-                ?: emptyList()
-
-            else -> error("No response")
-        }
-    }
-
-}
