+++ src/compiler/core/src/commonTest/kotlin/community/flock/wirespec/compiler/core/CompileRefinedTest.kt
@@ -28,6 +28,7 @@ class CompileRefinedTest {
             package community.flock.wirespec.generated
             
             import community.flock.wirespec.Wirespec
+            import kotlin.reflect.typeOf
 
             data class TodoId(override val value: String): Wirespec.Refined {
               override fun toString() = value
